[Fix][Connector-V2] update
