+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-jdbc-e2e/connector-jdbc-e2e-common/src/test/java/org/apache/seatunnel/connectors/seatunnel/jdbc/AbstractJdbcIT.java
@@ -31,6 +31,10 @@ import org.apache.seatunnel.common.exception.SeaTunnelRuntimeException;
 import org.apache.seatunnel.common.utils.ExceptionUtils;
 import org.apache.seatunnel.connectors.seatunnel.jdbc.catalog.iris.IrisCatalog;
 import org.apache.seatunnel.connectors.seatunnel.jdbc.catalog.oracle.OracleCatalog;
+import org.apache.seatunnel.connectors.seatunnel.jdbc.config.JdbcConnectionConfig;
+import org.apache.seatunnel.connectors.seatunnel.jdbc.config.JdbcSourceTableConfig;
+import org.apache.seatunnel.connectors.seatunnel.jdbc.source.JdbcSourceTable;
+import org.apache.seatunnel.connectors.seatunnel.jdbc.utils.JdbcCatalogUtils;
 import org.apache.seatunnel.e2e.common.TestResource;
 import org.apache.seatunnel.e2e.common.TestSuiteBase;
 import org.apache.seatunnel.e2e.common.container.ContainerExtendedFactory;
@@ -53,6 +57,7 @@ import org.testcontainers.images.PullPolicy;
 import org.testcontainers.lifecycle.Startables;
 
 import com.github.dockerjava.api.model.Image;
+import lombok.SneakyThrows;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -72,7 +77,9 @@ import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.Properties;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -321,7 +328,7 @@ public abstract class AbstractJdbcIT extends TestSuiteBase implements TestResour
                     dbServer.getDockerImageName(),
                     images);
             try {
-                dockerClient.removeImageCmd(dbServer.getDockerImageName()).exec();
+                // dockerClient.removeImageCmd(dbServer.getDockerImageName()).exec();
             } catch (Exception ignored) {
                 log.warn("Failed to delete the image. Another container may be in use", ignored);
             }
@@ -452,6 +459,43 @@ public abstract class AbstractJdbcIT extends TestSuiteBase implements TestResour
             catalog.dropDatabase(targetTablePath, false);
             Assertions.assertFalse(catalog.databaseExists(targetTablePath.getDatabaseName()));
         }
+    }
+
+    @SneakyThrows
+    @Test
+    public void testCatalogWithCatalogUtils() {
+        if (StringUtils.isBlank(jdbcCase.getTablePathFullName())) {
+            return;
+        }
+
+        List<JdbcSourceTableConfig> tablesConfig = new ArrayList<>();
+        JdbcSourceTableConfig tableConfig =
+                JdbcSourceTableConfig.builder()
+                        .query("SELECT * FROM " + jdbcCase.getSourceTable())
+                        .useSelectCount(false)
+                        .build();
+        tablesConfig.add(tableConfig);
+        Map<TablePath, JdbcSourceTable> tables =
+                JdbcCatalogUtils.getTables(
+                        JdbcConnectionConfig.builder()
+                                .url(jdbcCase.getJdbcUrl().replace(HOST, dbServer.getHost()))
+                                .username(jdbcCase.getUserName())
+                                .password(jdbcCase.getPassword())
+                                .build(),
+                        tablesConfig);
+        Set<TablePath> tablePaths = tables.keySet();
+
+        tablePaths.forEach(
+                tablePath -> {
+                    log.info(
+                            "Expected: {} Actual: {}",
+                            tablePath.getFullName(),
+                            jdbcCase.getTablePathFullName());
+                    Assertions.assertTrue(
+                            tablePath
+                                    .getFullName()
+                                    .equalsIgnoreCase(jdbcCase.getTablePathFullName()));
+                });
     }
 
     protected Object[] toArrayResult(ResultSet resultSet, String[] fieldNames)
