+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-jdbc-e2e/connector-jdbc-e2e-common/src/test/java/org/apache/seatunnel/connectors/seatunnel/jdbc/JdbcCase.java
@@ -57,4 +57,8 @@ public class JdbcCase {
     private String catalogDatabase;
     private String catalogSchema;
     private String catalogTable;
+
+    // The full path of the table created when initializing data
+    // According to whether jdbc api supports setting
+    private String tablePathFullName;
 }
