+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-jdbc-e2e/connector-jdbc-e2e-part-1/src/test/java/org/apache/seatunnel/connectors/seatunnel/jdbc/JdbcOracleIT.java
@@ -184,6 +184,8 @@ public class JdbcOracleIT extends AbstractJdbcIT {
                 .configFile(CONFIG_FILE)
                 .insertSql(insertSql)
                 .testData(testDataSet)
+                // oracle jdbc not support getTables/getCatalog/getSchema , is empty
+                .tablePathFullName(TablePath.DEFAULT.getFullName())
                 .build();
     }
